redesigned laneswitch:
* the entire FSM within the `laneswitches` module has been removed.
* It was observed that the consumer can request the data immediately
after closing the transfer from producer. This means that the `ceX` signals
of the memcores stay asserted all throughout the usage of the producer and
consumer. This meant that there is only a margin of 1 clock cycle available
to swich the lane.
* Since requests to the buffer are mutually exclusive (n_sections=1), there
will be no case where `laneswitch` instances need to be switched
independently. Moreover, the `laneswitch` RTL is much easier to edit than the
AST-generated `laneswitches` module. After this update `laneswitches` acts
only as a wrapper around the individual `laneswitch` instances.