[WIP] relaying for hybrid-buffers
* hierarchy of buffer has to change
* buffer:
  |---occupied_sections
  |---free_sections
  |---laneswitches
  `---memcores
* relay_buffer:
  |---relay_occupied_sections
  |---relay_free_sections
  |---relay_ls_memcores
      |---relay_memcores_reg
      `---ls_memcores
          |---laneswitches
          `---memcore

Since the laneswitches are being pulled in so deep, the combinational
path that monitors the `fifo_read`s will be affected.
`fifo_free_sections_read` from the producer (on lane 0) is flopped,
while `fifo_occupied_sections_read` from the consumer (on lane 1) is
read directly as a wire into the ls_memcores boundary.